Develop with nix flakes, upgrade to Veilid 0.3.2
Unable to source veilid-core 0.3.2 dependency through crates.io
directly; the build scripts fail.

Veilid 0.3.2 requires newer versions of capnproto and protobuf tooling
to be installed in order to compile, so this adds a Nix flake which
sources those dependencies into a `nix develop` shell.

Veilid 0.3.2 sources are themselves nix-packaged and vendored by
a symlink to the nix derivation.

This vendored link is easily replaced in a containerized Debian-based
build, when we're ready for Docker and binary builds for other platforms.